Read preferences as scripts common method
